:wrench: Merge Atmosphere and AtmosphereRenderer classes
