+++ plugins/csp-atmospheres/src/Atmosphere.hpp
@@ -9,47 +9,114 @@
 #define CSP_ATMOSPHERES_ATMOSPHERE_HPP
 
 #include "../../../src/cs-scene/CelestialObject.hpp"
-#include "AtmosphereRenderer.hpp"
+#include "ModelBase.hpp"
 #include "Plugin.hpp"
+
+#include <VistaBase/VistaVectorMath.h>
+#include <VistaKernel/DisplayManager/VistaViewport.h>
+#include <VistaKernel/GraphicsManager/VistaOpenGLDraw.h>
+#include <VistaOGLExt/VistaBufferObject.h>
+#include <VistaOGLExt/VistaGLSLShader.h>
+#include <VistaOGLExt/VistaTexture.h>
+#include <VistaOGLExt/VistaVertexArrayObject.h>
+
+#include <glm/glm.hpp>
+#include <memory>
+#include <unordered_map>
 
 namespace cs::core {
 class SolarSystem;
-}
+class GraphicsEngine;
+class EclipseShadowReceiver;
+} // namespace cs::core
+
+namespace cs::graphics {
+class ShadowMap;
+class HDRBuffer;
+} // namespace cs::graphics
 
 namespace csp::atmospheres {
 
-/// This is a wrapper around a AtmosphereRenderer, adding SPICE based positioning.
-class Atmosphere {
+/// This class draws a configurable atmosphere. Just put an OpenGLNode into your SceneGraph at the
+/// very same position as your planet. Set its scale to the same size as your planet.
+class Atmosphere : public IVistaOpenGLDraw {
  public:
-  Atmosphere(std::shared_ptr<Plugin::Settings> pluginSettings,
-      std::shared_ptr<cs::core::Settings>      settings,
-      std::shared_ptr<cs::core::SolarSystem> solarSystem, std::string objectName);
+  explicit Atmosphere(std::shared_ptr<Plugin::Settings> pluginSettings,
+      std::shared_ptr<cs::core::Settings>               allSettings,
+      std::shared_ptr<cs::core::SolarSystem>            solarSystem,
+      std::shared_ptr<cs::core::GraphicsEngine> graphicsEngine, std::string objectName);
+
   ~Atmosphere();
 
-  Atmosphere(Atmosphere const& other) = delete;
-  Atmosphere(Atmosphere&& other)      = delete;
-
-  Atmosphere& operator=(Atmosphere const& other) = delete;
-  Atmosphere& operator=(Atmosphere&& other) = delete;
-
-  /// Configures the internal renderer according to the given values.
   void configure(Plugin::Settings::Atmosphere const& settings);
 
-  /// Access the internal AtmosphereRender to configure atmosphere parameters.
-  AtmosphereRenderer&       getRenderer();
-  AtmosphereRenderer const& getRenderer() const;
-
-  /// This is called once a frame by the plugin. It updates the EclipseShadowReceiver.
   void update();
 
+  bool Do() override;
+  bool GetBoundingBox(VistaBoundingBox& bb) override;
+
  private:
-  std::shared_ptr<Plugin::Settings>                mPluginSettings;
-  std::shared_ptr<cs::core::Settings>              mAllSettings;
-  std::shared_ptr<cs::core::SolarSystem>           mSolarSystem;
-  std::string                                      mObjectName;
+  void updateShader();
+
+  std::shared_ptr<Plugin::Settings>         mPluginSettings;
+  std::shared_ptr<cs::core::Settings>       mAllSettings;
+  std::shared_ptr<cs::core::SolarSystem>    mSolarSystem;
+  std::shared_ptr<cs::core::GraphicsEngine> mGraphicsEngine;
+  std::string                               mObjectName;
+  std::unique_ptr<VistaOpenGLNode>          mAtmosphereNode;
+
+  // std::unique_ptr<VistaTexture>     mCloudTexture;
+  // std::string                       mCloudTextureFile;
+  // float                             mCloudHeight    = 0.001F;
+  // bool                              mUseClouds      = false;
+
+  glm::dvec3                   mRadii          = glm::dvec3(1.0, 1.0, 1.0);
+  glm::dmat4                   mWorldTransform = glm::dmat4(1.0);
+  Plugin::Settings::Atmosphere mSettings;
+
+  int mEnableShadowsConnection = -1;
+  int mEnableHDRConnection     = -1;
+
+  std::shared_ptr<cs::graphics::ShadowMap>         mShadowMap;
+  std::shared_ptr<cs::graphics::HDRBuffer>         mHDRBuffer;
   std::shared_ptr<cs::core::EclipseShadowReceiver> mEclipseShadowReceiver;
-  AtmosphereRenderer                               mRenderer;
-  std::unique_ptr<VistaOpenGLNode>                 mAtmosphereNode;
+
+  VistaGLSLShader        mAtmoShader;
+  VistaVertexArrayObject mQuadVAO;
+  VistaBufferObject      mQuadVBO;
+
+  struct GBufferData {
+    std::unique_ptr<VistaTexture> mDepthBuffer;
+    std::unique_ptr<VistaTexture> mColorBuffer;
+  };
+
+  std::unordered_map<VistaViewport*, GBufferData> mGBufferData;
+
+  bool      mShaderDirty    = true;
+  float     mSunIlluminance = 1.F;
+  glm::vec3 mSunDirection   = glm::vec3(1, 0, 0);
+
+  struct {
+    uint32_t sunIntensity   = 0;
+    uint32_t sunDir         = 0;
+    uint32_t depthBuffer    = 0;
+    uint32_t colorBuffer    = 0;
+    uint32_t cloudTexture   = 0;
+    uint32_t cloudAltitude  = 0;
+    uint32_t shadowCascades = 0;
+    uint32_t sunIlluminance = 0;
+
+    std::array<uint32_t, 5> shadowMaps{};
+    std::array<uint32_t, 5> shadowProjectionMatrices{};
+
+    uint32_t inverseModelViewMatrix           = 0;
+    uint32_t inverseModelViewProjectionMatrix = 0;
+    uint32_t inverseProjectionMatrix          = 0;
+    uint32_t modelViewMatrix                  = 0;
+    uint32_t modelMatrix                      = 0;
+  } mUniforms;
+
+  std::unique_ptr<ModelBase> mModel;
 };
 
 } // namespace csp::atmospheres
