+++ plugins/csp-atmospheres/src/Plugin.hpp
@@ -30,20 +30,21 @@ class Plugin : public cs::core::PluginBase {
       double                            mHeight; ///< In meters.
       cs::utils::DefaultProperty<Model> mModel{Model::eCosmoScoutVR};
       nlohmann::json                    mModelSettings;
-
-      cs::utils::DefaultProperty<bool>  mEnable{true};
       cs::utils::DefaultProperty<bool>  mEnableWater{false};
       cs::utils::DefaultProperty<float> mWaterLevel{0.F}; ///< In meters.
       cs::utils::DefaultProperty<bool>  mEnableClouds{true};
       std::optional<std::string>        mCloudTexture;          ///< Path to the cloud texture.
       cs::utils::DefaultProperty<float> mCloudAltitude{3000.F}; ///< In meters.
-      cs::utils::DefaultProperty<bool>  mEnableLightShafts{false};
 
       bool operator==(Atmosphere const& other) const;
       bool operator!=(Atmosphere const& other) const;
     };
 
     std::unordered_map<std::string, Atmosphere> mAtmospheres;
+
+    cs::utils::DefaultProperty<bool> mEnable{true};
+    cs::utils::DefaultProperty<bool> mEnableLightShafts{false};
+    cs::utils::DefaultProperty<bool> mEnableIndirectLighting{false};
   };
 
   void init() override;
@@ -59,11 +60,9 @@ class Plugin : public cs::core::PluginBase {
   std::unordered_map<std::string, std::shared_ptr<Atmosphere>> mAtmospheres;
   std::string                                                  mActiveAtmosphere;
 
-  int mActiveObjectConnection  = -1;
-  int mEnableShadowsConnection = -1;
-  int mEnableHDRConnection     = -1;
-  int mOnLoadConnection        = -1;
-  int mOnSaveConnection        = -1;
+  int mActiveObjectConnection = -1;
+  int mOnLoadConnection       = -1;
+  int mOnSaveConnection       = -1;
 };
 
 } // namespace csp::atmospheres
